[295437] fixed: XQDT SF BUG#2883207: two dashes in an element
No satisfied with the solution. The lexer switcing has to be improved: now the lexer is digesting more tokens that it needs after a context switch. Therefore false errors are reported. In this fix, those false errors are hidden from the UI by implementing a delayed error reporting in lexers.